First `expand_dependencies` logic edge case
In the case of running under jruby, on a frozen bundle, and against a
lockfile only including the ruby platform, `expand_dependencies` would
end up filtering out all dependencies, causing errors.

The reason is that on a frozen bundle, the local platform is not added
to the array of definition platforms, since only platform information
from the lockfile should be used. That means that under the previous
logic, `expand_dependencies` would filter out everything.

The solution is to make sure to always _at least_ expand all
dependencies with the local platform.